Book display bugs & faster book fetch

+++ api/business-logic/users.js
@@ -1,5 +1,4 @@
 const dataAccess = require('../data-access/mongodbAccess');
-const googleManager = require('./googleBooksAPI');
 const ObjectID = require('mongodb').ObjectID;
 
 const usersStore = dataAccess('Users');
@@ -63,13 +62,9 @@ const usersManager = {
       searchedBooks.push(bookWithFullAccess);
     });
 
-    // console.log(searchedBooks);
-
     for (let index = 0; index < searchedBooks.length; index++) {
       const book = searchedBooks[index];
       let user = '';
-
-      // console.log(book.book_userId);
 
       // add city
       try {
@@ -77,28 +72,12 @@ const usersManager = {
         user = await usersStore.getById(book.book_userId);
         book.city = user.city;
         book.username = user.username;
+        book.userEmail = user.email;
       } catch (error) {
         console.log('cannot get city for book ' + book.title, error);
       }
-
-      // add picture url
-
-      try {
-        /* TODO to improve performance frontend should not make a call to google
-         * while loading books first it should load from our api and show the book data
-         * then make a call to google to get the picture url and then rerender all
-         */
-        const thumbnail = await googleManager.getPictureURL(book.isbn_10);
-        book.thumbnail = thumbnail;
-        console.log('manager', thumbnail);
-      } catch (error) {
-        console.log('cannot get thumbnail for book ' + book.title, error);
-      }
     }
 
-    // TODO complete to 30 from google api
-
-    console.log(searchedBooks);
     return searchedBooks;
   },
   searchBookById: async (bookId) => {
@@ -128,29 +107,13 @@ const usersManager = {
       console.log('cannot get city for book ' + book.title, error);
     }
 
-    // add picture url
-
-    try {
-      /* TODO to improve performance frontend should not make a call to google
-       * while loading books first it should load from our api and show the book data
-       * then make a call to google to get the picture url and then rerender all
-       */
-      const thumbnail = await googleManager.getPictureURL(book.isbn_10);
-      bookWithFullAccess.thumbnail = thumbnail;
-    } catch (error) {
-      console.log('cannot get thumbnail for book ' + book.title, error);
-    }
-
     return bookWithFullAccess;
   },
   addUserBook: async (bookDetails) => {
     bookDetails.userId = new ObjectID(bookDetails.userId);
 
-    // bookDetails.userId = { $oid: `${bookDetails.userId}` };
-    console.log('object');
     const response = await booksStore.insert(bookDetails);
 
-    console.log(response);
     if (response.acknowledged) {
       return { message: 'Book added' };
     }
@@ -165,8 +128,14 @@ const usersManager = {
       throw new Error('Book not updated');
     }
   },
-  deleteUserBook: async (bookId) => {
-    const response = await booksStore.delete(bookId);
+  deleteUserBook: async (bookId, userId) => {
+    const book = await booksStore.getById(bookId);
+
+    let response;
+
+    if (book.userId.toString() === userId) {
+      response = await booksStore.delete(bookId);
+    }
 
     console.log(response);
